pyproject.toml
[tool.poetry]
name = "doter"
version = "1.0.1"
description = "A dotfile manager"
authors = ["Alex Zhang <zhangchi0104@live.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
rich = "^10.15.2"
pydantic = "^1.10.4"
pyyaml = "^6.0"
joblib = "^1.2.0"

[tool.poetry.dev-dependencies]
toml = "^0.10.2"
yapf = "^0.31.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"
pytest-asyncio = "^0.20.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
doter = "doter.__main__:main"
